// src/generated-api/models/RSSItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BMI-LAB Web API
 * BMI-LAB을 관리하기 위한 웹 서비스
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RSSItem
 */
export interface RSSItem {
    /**
     * 
     * @type {string}
     * @memberof RSSItem
     */
    author?: string;
    /**
     * 
     * @type {Date}
     * @memberof RSSItem
     */
    startDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof RSSItem
     */
    endDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof RSSItem
     */
    link?: string;
    /**
     * 
     * @type {string}
     * @memberof RSSItem
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof RSSItem
     */
    category?: string;
    /**
     * 
     * @type {Date}
     * @memberof RSSItem
     */
    publishedAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof RSSItem
     */
    budget?: number;
}

/**
 * Check if a given object implements the RSSItem interface.
 */
export function instanceOfRSSItem(value: object): value is RSSItem {
    return true;
}

export function RSSItemFromJSON(json: any): RSSItem {
    return RSSItemFromJSONTyped(json, false);
}

export function RSSItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): RSSItem {
    if (json == null) {
        return json;
    }
    return {
        
        'author': json['author'] == null ? undefined : json['author'],
        'startDate': json['startDate'] == null ? undefined : (new Date(json['startDate'])),
        'endDate': json['endDate'] == null ? undefined : (new Date(json['endDate'])),
        'link': json['link'] == null ? undefined : json['link'],
        'title': json['title'] == null ? undefined : json['title'],
        'category': json['category'] == null ? undefined : json['category'],
        'publishedAt': json['publishedAt'] == null ? undefined : (new Date(json['publishedAt'])),
        'budget': json['budget'] == null ? undefined : json['budget'],
    };
}

export function RSSItemToJSON(json: any): RSSItem {
    return RSSItemToJSONTyped(json, false);
}

export function RSSItemToJSONTyped(value?: RSSItem | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'author': value['author'],
        'startDate': value['startDate'] == null ? undefined : ((value['startDate']).toISOString().substring(0,10)),
        'endDate': value['endDate'] == null ? undefined : ((value['endDate']).toISOString().substring(0,10)),
        'link': value['link'],
        'title': value['title'],
        'category': value['category'],
        'publishedAt': value['publishedAt'] == null ? undefined : ((value['publishedAt']).toISOString().substring(0,10)),
        'budget': value['budget'],
    };
}